+++ modules/constants.mjs
@@ -1,4 +1,7 @@
 let constants = {
-  moduleName: "forien-unidentified-items"
+  modulePath: "modules/forien-unidentified-items",
+  moduleName: "forien-unidentified-items",
+  defaultIcon: "unidentified.png",
+  moduleLabel: "Forien's Unidentified Items"
 };
 export default constants;